refactor: move encoding pkg
