+++ pkg/encoding/error.go
