+++ pkg/primitive/binary.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 
 	"github.com/pkg/errors"
-	encoding2 "github.com/siyul-park/uniflow/internal/encoding"
+	encoding2 "github.com/siyul-park/uniflow/pkg/encoding"
 )
 
 type (
