+++ pkg/primitive/shortcut.go
@@ -2,7 +2,7 @@ package primitive
 
 import (
 	"github.com/pkg/errors"
-	"github.com/siyul-park/uniflow/internal/encoding"
+	"github.com/siyul-park/uniflow/pkg/encoding"
 )
 
 // NewPointerEncoder is encode Object to Object.
