+++ pkg/primitive/slice.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/benbjohnson/immutable"
 	"github.com/pkg/errors"
-	"github.com/siyul-park/uniflow/internal/encoding"
+	"github.com/siyul-park/uniflow/pkg/encoding"
 )
 
 type (
