+++ pkg/scheme/codec.go
@@ -2,7 +2,7 @@ package scheme
 
 import (
 	"github.com/pkg/errors"
-	"github.com/siyul-park/uniflow/internal/encoding"
+	"github.com/siyul-park/uniflow/pkg/encoding"
 	"github.com/siyul-park/uniflow/pkg/node"
 )
